Created map that maps steps to moves and returns a JS object / React component for it
"For each step in the history, we create a list item <li> with a button <button> inside it that has a click handler which we’ll implement shortly. With this code, you should see a list of the moves that have been made in the game, along with a warning that says:

    Warning: Each child in an array or iterator should have a unique “key” prop. Check the render method of “Game”."

"Let’s talk about what that warning means."

"index.js:2178 Warning: Each child in an array or iterator should have a unique "key" prop.

Check the render method of `Game`. See https://fb.me/react-warning-keys for more information.
    in li (at index.js:185)
    in Game (at index.js:246)"